requirements.txt
google-generativeai
edge-tts
google-api-python-client
google-auth-oauthlib
requests
pillow